split the EventHandlerGolem in two and create EventHandlerGolemServer
